Revert "Crawl seperate "
