--- src/app/menu/Component/modal.module.css
@@ -1,49 +1,0 @@
-.modal_container {
-  position: relative;
-  width: 100%;
-  background-color: #fff;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal_container button.close {
-  position: absolute;
-  top: 1rem;
-  right: 1rem;
-}
-
-.modal_container form {
-  width: 80%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.modal_container form input {
-  height: 2rem;
-  width: 100%;
-}
-.modal_container form button {
-  margin-top: 1rem;
-  height: 2rem;
-}
-/** 메뉴 선택 모달 */
-.modal_container .menu-column {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  margin-bottom: 1rem;
-}
-
-.modal_container .menu-column label {
-  font-size: 1.5rem;
-  margin-bottom: 0.5rem;
-}
-.modal_container .menu-column select {
-  width: 100%;
-  height: 2rem;
-}
