+++ src/crawling/index.ts
@@ -1,77 +1,57 @@
 import { CheerioAPI, load } from "cheerio";
-import { MenuProps, ingredientList, initialIngredient } from "@/type";
+import { MenuContentsProps, ingredientList, initialIngredient } from "@/type";
 
-export async function getCategories(url: string) {
+export async function getMenu(url: string) {
+  const urlObject = new URL(url);
+  const protocolAndHostname = `${urlObject.protocol}//${urlObject.hostname}`;
   const coffeeBeanPages = await fetch(url, { cache: "no-store" })
     .then((res) => res.text())
     .then(load)
     .then(getHrefFromTags)
-    .then((list) =>
-      list.map(({ title, category }) => {
-        const [, categoryNum] = category.split("?category=");
-        return {
-          title,
-          category: categoryNum,
-        };
-      })
-    );
-  return coffeeBeanPages;
+    .then((list) => list.map((href) => `${protocolAndHostname}${href}`));
+
+  const menuContent: MenuContentsProps[] = [];
+  for await (const page of coffeeBeanPages) {
+    await new Promise((res) => setTimeout(res, 500));
+    const menu = await getMenuFromPages(page, protocolAndHostname);
+    menuContent.push(menu);
+  }
+
+  return menuContent;
 }
 
 function getHrefFromTags($: CheerioAPI) {
   const contents = $("#contents .lnb_wrap2 li:first-child a").toArray();
-  const namesAndCategories = contents
-    .map((el) => {
-      const $this = $(el);
-      return {
-        title: $this.html() || "",
-        category: $this.attr("href") || "",
-      };
-    })
-    .filter(
-      ({ title, category }) =>
-        category &&
-        category?.includes("category") &&
-        title &&
-        title.localeCompare("음료") !== 0
-    );
-
-  const result = namesAndCategories.reduce<typeof namesAndCategories>(
-    (acc, item) =>
-      acc.find(({ category }) => category === item.category)
-        ? acc
-        : [...acc, item],
-    []
-  );
-
-  return result;
+  const categories = contents
+    .map((el) => $(el).attr("href") || "")
+    .filter((href) => href && href?.includes("category"));
+  return [...new Set(categories)];
 }
 
-export async function getMenuFromPages(url: string) {
-  let result: MenuProps[] = [];
-  const urlObject = new URL(url);
-  const protocolAndHostname = `${urlObject.protocol}//${urlObject.hostname}`;
+async function getMenuFromPages(page: string, protocolAndHostname: string) {
+  const result: MenuContentsProps = { title: "", list: [] };
   let cnt = 1;
   while (cnt < 6) {
     await new Promise((res) => setTimeout(res, 500));
-    const data = await fetch(`${url}&page=${cnt++}`, { cache: "no-store" })
+    const data = await fetch(`${page}&page=${cnt++}`, { cache: "no-store" })
       .then((res) => res.text())
       .then(load)
       .then(getMenuFromPage(protocolAndHostname));
-    result = [...result, ...data.list];
+    if (!result.title) {
+      result.title = data.title;
+    }
+    result.list = [...result.list, ...data.list];
     if (!data.hasNextPage) break;
   }
   return result;
 }
 
 function getMenuFromPage(protocolAndHostname: string) {
-  return function ($: CheerioAPI): { list: MenuProps[]; hasNextPage: boolean } {
-    const breadcrumb = $("#wrap .location li a");
-    const href = breadcrumb.last().attr("href");
-    let category: string;
-    if (href) {
-      [, category] = href.split("=");
-    }
+  return function (
+    $: CheerioAPI
+  ): MenuContentsProps & { hasNextPage: boolean } {
+    const breadcrumb = $("#wrap .location li");
+    const title = breadcrumb.last().text();
     const contents = $("#contents .menu_list li").toArray();
     const list = contents.map((el) => {
       const $this = $(el);
@@ -99,7 +79,6 @@ function getMenuFromPage(protocolAndHostname: string) {
         photo,
         name,
         description,
-        category,
         info,
       };
     });
@@ -107,6 +86,7 @@ function getMenuFromPage(protocolAndHostname: string) {
     const hasNextPage = pagination.some((el) => el.attribs.class === "next");
 
     return {
+      title,
       list,
       hasNextPage,
     };
