+++ src/crawling/mock.ts
@@ -1,60 +1,52 @@
-import { Category, initialIngredient } from "../type";
+import { MenuContentsProps, initialIngredient } from "../type";
 
-const CATEGORY_LIST: Array<Category> = [
+const MENULIST: MenuContentsProps[] = [
   {
     title: "커피",
-    category: "13",
+    list: [
+      { kor: "아메리카노", eng: "Americano" },
+      { kor: "카페라떼", eng: "Cafe Latte" },
+      { kor: "카페모카", eng: "Cafe Moca" },
+      { kor: "콜드브루", eng: "Cold Brew" },
+    ].map((name) => ({
+      photo: "none",
+      name,
+      description: name.kor,
+      info: initialIngredient,
+    })),
   },
   {
     title: "티",
-    category: "18",
+    list: [
+      {
+        kor: "레모네이드",
+        eng: "Lemonade",
+      },
+      {
+        kor: "캐모마일",
+        eng: "chamomile",
+      },
+      {
+        kor: "히비스커스",
+        eng: "Hibiscus",
+      },
+      {
+        kor: "녹차",
+        eng: "Green tea",
+      },
+      {
+        kor: "아샷추",
+        eng: "Icetea with shot",
+      },
+    ].map((name) => ({
+      photo: "none",
+      name,
+      description: name.kor,
+      info: initialIngredient,
+    })),
   },
 ];
 
-const MENULIST = [
-  ...[
-    { kor: "아메리카노", eng: "Americano" },
-    { kor: "카페라떼", eng: "Cafe Latte" },
-    { kor: "카페모카", eng: "Cafe Moca" },
-    { kor: "콜드브루", eng: "Cold Brew" },
-  ].map((name) => ({
-    photo: "none",
-    name,
-    description: name.kor,
-    info: initialIngredient,
-    category: "13",
-  })),
-  ...[
-    {
-      kor: "레모네이드",
-      eng: "Lemonade",
-    },
-    {
-      kor: "캐모마일",
-      eng: "chamomile",
-    },
-    {
-      kor: "히비스커스",
-      eng: "Hibiscus",
-    },
-    {
-      kor: "녹차",
-      eng: "Green tea",
-    },
-    {
-      kor: "아샷추",
-      eng: "Icetea with shot",
-    },
-  ].map((name) => ({
-    photo: "none",
-    name,
-    description: name.kor,
-    info: initialIngredient,
-    category: "18",
-  })),
-];
-
 export const MOCK = {
-  CATEGORY_LIST,
   MENULIST,
 } as const;
