+++ src/database/coffeebean/index.ts
@@ -1,7 +1,6 @@
 export const COFFEEBEAN = {
   DB_NAME: "coffeebean",
   COLLECTION: {
-    CATEGORY: "category",
     ORDER: "order",
     MENU: "menu",
   },
